+++ components/ui/icons.tsx
@@ -235,6 +235,15 @@ export function RotateCcw(props: IconProps) {
   )
 }
 
+export function Smartphone(props: IconProps) {
+  return (
+    <BaseIcon {...props}>
+      <rect width={"14"} height={"20"} x={"5"} y={"2"} rx={"2"} ry={"2"} />
+      <path d={"M12 18h.01"} />
+    </BaseIcon>
+  )
+}
+
 export function Search(props: IconProps) {
   return (
     <BaseIcon {...props}>